platform/x86: Update surface SAM driver
Update Surface System Aggregator Module driver based on module.

Changes:
- Enable performance modes on Surface Laptop 1 and 2.
- Support for D0-exit/D0-entry notifications, sent to the SAM controller
  when suspending (D0-exit) or resuming (D0-entry) on newer (gen. 7+)
  devices.
- Replace sysfs-based debug device with new debugfs-based debug device.
- More internal restructuring and internal API changes.
- Some small bug fixes.

Links:
- SAM: https://github.com/linux-surface/surface-aggregator-module/commit/89b8232d49856a3a1e9710761a87d81c347b574f

+++ drivers/platform/x86/surface_sam/surface_sam_san.h
@@ -18,7 +18,7 @@ struct surface_sam_san_rqsg {
 	u8 tc;				// target category
 	u8 cid;				// command ID
 	u8 iid;				// instance ID
-	u8 cdl;				// command data length (length of payload)
+	u16 cdl;			// command data length (length of payload)
 	u8 *pld;			// pointer to payload of length cdl
 };
